feat(render-model): swap agent chat to RenderModel + Adapter path
Move agent chat build/encode logic from gui.ex to the RenderModel
pattern: model (pre-encoded binary + fingerprint), builder (handles
message styling, prompt completion, help groups), adapter encoder
(fingerprint-based skip). Uses :not_visible sentinel. The builder
gracefully handles missing shell_state by catching errors from
active_session calls.
