+++ lib/minga_editor/frontend/emit/gui.ex
@@ -19,9 +19,6 @@ defmodule MingaEditor.Frontend.Emit.GUI do
   Called from `Emit.emit/3` only when the frontend has GUI capabilities.
   """
 
-  alias MingaAgent.Session, as: AgentSession
-  alias MingaEditor.Agent.UIState
-  alias MingaEditor.Agent.View.PromptSemanticWindow
   alias Minga.Buffer
   alias Minga.Config
 
@@ -123,7 +120,6 @@ defmodule MingaEditor.Frontend.Emit.GUI do
     # Each build_gui_* function returns {cmd | nil, updated_caches}.
     # Note: status_bar is now handled by the RenderModel adapter path.
     builders = [
-      &build_gui_agent_chat_cmd/2,
       &build_gui_hover_popup_cmd/2,
       &build_gui_float_popup_cmd/2,
       &build_gui_change_summary_cmd/2,
@@ -160,230 +156,6 @@ defmodule MingaEditor.Frontend.Emit.GUI do
       buf -> Buffer.file_path(buf)
     end
   end
-
-  # ── Agent chat ──
-
-  @spec build_gui_agent_chat_cmd(ctx(), Caches.t()) :: {binary() | nil, Caches.t()}
-  defp build_gui_agent_chat_cmd(ctx, caches) do
-    active_window = Map.get(ctx.windows.map, ctx.windows.active)
-    is_agent_chat = active_window != nil && Content.agent_chat?(active_window.content)
-    session = ctx.shell.active_session(ctx.shell_state)
-
-    # Compute fingerprint from cheap state fields to avoid calling
-    # AgentSession.messages (expensive GenServer.call that allocates a
-    # formatted message list) on every frame. The prompt buffer content
-    # is included because it's a single fast GenServer.call, and the PID
-    # alone is stable (wouldn't detect typing). The styled cache length
-    # is a reliable proxy for message count changes because styling
-    # happens in the same render cycle as message arrival.
-    # message_version is bumped on every :messages_changed event,
-    # including collapse toggles, ensuring the fingerprint changes.
-    {fp, prompt_text} =
-      if is_agent_chat && session do
-        panel = ctx.agent_ui.panel
-        view = ctx.agent_ui.view
-        styled_len = length(panel.cached_styled_messages || [])
-        text = safe_prompt_content(panel.prompt_buffer)
-        prompt_cursor = UIState.input_cursor(panel)
-        prompt_line_count = UIState.input_line_count(panel)
-        inner_width = max(ctx.viewport.cols - 10, 20)
-        visible_rows = PromptSemanticWindow.visible_rows(panel, inner_width)
-
-        {:erlang.phash2(
-           {:visible, ctx.shell_state.agent.runtime.status,
-            ctx.shell_state.agent.pending_approval, styled_len, panel.model_name,
-            panel.thinking_level, text, panel.message_version, view.help_visible, view.focus,
-            ctx.editing.mode, prompt_cursor, prompt_line_count, visible_rows,
-            panel.mention_completion}
-         ), text}
-      else
-        {:not_visible, ""}
-      end
-
-    if fp != caches.last_gui_agent_chat_fp do
-      data = build_agent_chat_data(ctx, prompt_text)
-
-      if data.visible do
-        log_agent_chat_message_stats(data.messages)
-      end
-
-      {ProtocolGUI.encode_gui_agent_chat(data), %{caches | last_gui_agent_chat_fp: fp}}
-    else
-      {nil, caches}
-    end
-  end
-
-  @spec log_agent_chat_message_stats([{pos_integer(), term()}]) :: :ok
-  defp log_agent_chat_message_stats(messages) do
-    {styled, plain} =
-      Enum.reduce(messages, {0, 0}, fn
-        {_, {:styled_assistant, _}}, {s, p} -> {s + 1, p}
-        {_, {:styled_tool_call, _, _}}, {s, p} -> {s + 1, p}
-        {_, {:assistant, _}}, {s, p} -> {s, p + 1}
-        _, acc -> acc
-      end)
-
-    Minga.Log.debug(
-      :render,
-      "[gui] sending agent chat: #{length(messages)} msgs (#{styled} styled, #{plain} plain assistant)"
-    )
-  end
-
-  # Builds the prompt completion popup data for @-mention or /slash completion.
-  # Returns a map suitable for encode_prompt_completion, or nil if no completion active.
-  @spec build_prompt_completion(MingaEditor.Agent.UIState.Panel.t()) :: map() | nil
-  defp build_prompt_completion(%{mention_completion: %{candidates: candidates} = comp})
-       when is_list(candidates) and candidates != [] do
-    # Slash completions carry {name, description} tuples in :slash_candidates.
-    # Mention completions are plain file path strings.
-    {type, formatted_candidates} =
-      case comp[:slash_candidates] do
-        slash when is_list(slash) and slash != [] ->
-          {:slash, slash}
-
-        _ ->
-          {:mention, candidates}
-      end
-
-    %{
-      type: type,
-      candidates: formatted_candidates,
-      selected: comp.selected,
-      anchor_line: comp.anchor_line,
-      anchor_col: comp.anchor_col
-    }
-  end
-
-  defp build_prompt_completion(_panel), do: nil
-
-  # Reads prompt buffer content, guarding against a dead process.
-  # The prompt buffer can die between state updates; the :DOWN handler
-  # clears the PID on the next cycle, but there's a race window.
-  @spec safe_prompt_content(pid() | nil) :: String.t()
-  defp safe_prompt_content(nil), do: ""
-
-  defp safe_prompt_content(buf) do
-    Buffer.content(buf) |> String.trim_trailing("\n")
-  catch
-    :exit, _ -> ""
-  end
-
-  @spec build_agent_chat_data(ctx(), String.t()) :: map()
-  defp build_agent_chat_data(ctx, prompt_text) do
-    active_window = Map.get(ctx.windows.map, ctx.windows.active)
-    is_agent_chat = active_window != nil && Content.agent_chat?(active_window.content)
-    session = ctx.shell.active_session(ctx.shell_state)
-
-    if is_agent_chat && session do
-      messages_with_ids =
-        try do
-          AgentSession.messages_with_ids(session)
-        catch
-          :exit, _ -> []
-        end
-
-      # Use cached styled runs for assistant messages when available.
-      # This avoids recomputing tree-sitter/markdown styling per frame.
-      styled_cache = ctx.agent_ui.panel.cached_styled_messages
-      pending_approval = ctx.shell_state.agent.pending_approval
-      gui_messages = build_gui_messages(messages_with_ids, styled_cache, pending_approval)
-
-      view = ctx.agent_ui.view
-      help_visible = view.help_visible
-
-      help_groups =
-        if help_visible do
-          Minga.Keymap.Scope.Agent.help_groups(view.focus)
-        else
-          []
-        end
-
-      panel = ctx.agent_ui.panel
-      {cursor_line, cursor_col} = UIState.input_cursor(panel)
-      vim_mode = ctx.editing.mode
-      inner_width = max(ctx.viewport.cols - 10, 20)
-      visible_rows = PromptSemanticWindow.visible_rows(panel, inner_width)
-      prompt_completion = build_prompt_completion(panel)
-
-      %{
-        visible: true,
-        messages: gui_messages,
-        status: ctx.shell_state.agent.runtime.status || :idle,
-        model: ctx.agent_ui.panel.model_name,
-        thinking_level: panel.thinking_level,
-        prompt: prompt_text,
-        pending_approval: nil,
-        help_visible: help_visible,
-        help_groups: help_groups,
-        prompt_line_count: UIState.input_line_count(panel),
-        prompt_cursor_line: cursor_line,
-        prompt_cursor_col: cursor_col,
-        prompt_vim_mode: vim_mode,
-        prompt_visible_rows: visible_rows,
-        prompt_completion: prompt_completion
-      }
-    else
-      %{visible: false}
-    end
-  end
-
-  # Builds the message list for GUI encoding. Each entry is `{id, gui_message}`
-  # where `id` is the stable BEAM-assigned message ID. Replaces assistant messages
-  # with {:styled_assistant, styled_lines} when cached styled runs are available.
-  #
-  # Input: list of `{id, Message.t()}` pairs from `messages_with_ids/1`.
-  # Uses Enum.zip (O(n)) instead of Enum.at in a loop (O(n²)) since this
-  # runs every render frame at 60fps.
-  @spec build_gui_messages([{pos_integer(), term()}], [term()] | nil, map() | nil) :: [
-          {pos_integer(), term()}
-        ]
-  defp build_gui_messages(messages_with_ids, nil, pending_approval) do
-    Enum.map(messages_with_ids, &maybe_inline_approval(&1, pending_approval))
-  end
-
-  defp build_gui_messages(messages_with_ids, styled_cache, pending_approval)
-       when is_list(styled_cache) do
-    # Pad the cache to match message length if needed (messages may have grown)
-    padded = pad_cache(styled_cache, length(messages_with_ids))
-
-    Enum.zip(messages_with_ids, padded)
-    |> Enum.map(&maybe_style_message(&1, pending_approval))
-  end
-
-  @spec maybe_style_message({{pos_integer(), term()}, term()}, map() | nil) ::
-          {pos_integer(), term()}
-  defp maybe_style_message({{id, {:assistant, _text} = msg}, nil}, _pending_approval),
-    do: {id, msg}
-
-  defp maybe_style_message({{id, {:assistant, _text}}, styled_lines}, _pending_approval),
-    do: {id, {:styled_assistant, styled_lines}}
-
-  defp maybe_style_message({{id, {:tool_call, tc} = msg}, styled_lines}, pending_approval) do
-    case maybe_inline_approval({id, msg}, pending_approval) do
-      {^id, {:approval_tool_call, _tc, _approval}} = approval_message ->
-        approval_message
-
-      {^id, {:tool_call, _tc}} when is_list(styled_lines) ->
-        {id, {:styled_tool_call, tc, styled_lines}}
-
-      unchanged ->
-        unchanged
-    end
-  end
-
-  defp maybe_style_message({{id, msg}, _cache_entry}, _pending_approval), do: {id, msg}
-
-  @spec maybe_inline_approval({pos_integer(), term()}, map() | nil) :: {pos_integer(), term()}
-  defp maybe_inline_approval({id, {:tool_call, tc}}, %{tool_call_id: tool_call_id} = approval)
-       when tc.id == tool_call_id do
-    {id, {:approval_tool_call, tc, approval}}
-  end
-
-  defp maybe_inline_approval({id, msg}, _pending_approval), do: {id, msg}
-
-  @spec pad_cache([term()], non_neg_integer()) :: [term()]
-  defp pad_cache(cache, target_len) when length(cache) >= target_len, do: cache
-  defp pad_cache(cache, target_len), do: cache ++ List.duplicate(nil, target_len - length(cache))
 
   # ── Gutter separator ──
